File iteration logic, and finish out db open code.

+++ bbdb.go
@@ -4,6 +4,8 @@ import (
 	"flag"
 	"log"
 	"os"
+	"path/filepath"
+	"strings"
 )
 
 const (
@@ -11,8 +13,9 @@ const (
 )
 
 var (
-	dbFile string
-	wipe bool
+	dbFile  string
+	wipe    bool
+	driveDB *db
 )
 
 func main() {
@@ -23,14 +26,47 @@ func main() {
 	flag.Parse()
 
 	if wipe {
-		err = os.Remove(dbFile)
-		if err != nil {
-			log.Fatal(err)
+		_, err := os.Stat(dbFile)
+		if err == nil {
+			err = os.Remove(dbFile)
+			if err != nil {
+				log.Fatal(err)
+			}
 		}
 	}
 
-	_, err = newDB(dbFile)
+	driveDB, err = newDB(dbFile)
 	if err != nil {
 		log.Fatal(err)
 	}
+	defer driveDB.close()
+
+	for _, dir := range flag.Args() {
+		readOneDir(dir)
+	}
+}
+
+func readOneDir(dir string) {
+	filepath.Walk(dir, func(path string, info os.FileInfo, inErr error) (outErr error) {
+		if info != nil &&
+			!info.IsDir() &&
+			strings.HasSuffix(strings.ToLower(info.Name()), ".csv") {
+			err := readOneFile(path)
+			if err != nil {
+				log.Println(err)
+			}
+		}
+		return
+	})
+}
+
+func readOneFile(fname string) (err error) {
+	log.Println(fname)
+	s, err := openFile(fname)
+	if err != nil {
+		return
+	}
+	defer s.Close()
+
+	return
 }
